Added listToDos test
