+++ src/main/kotlin/health/patchwork/Item.kt
@@ -1,0 +1,5 @@
+package health.patchwork
+
+class Item(name: String) {
+
+}
