+++ src/main/kotlin/health/patchwork/Main.kt
@@ -1,3 +1,5 @@
+package health.patchwork
+
 import org.http4k.core.HttpHandler
 import org.http4k.core.Response
 import org.http4k.core.Status
@@ -5,9 +7,7 @@ import org.http4k.server.Undertow
 import org.http4k.server.asServer
 
 fun main() {
-
     handler.asServer(Undertow(8080)).start()
-
-
 }
+
 val handler : HttpHandler = { request -> Response(Status.OK).body("hello") }
